test(cli): split passthrough suites
