+++ packages/cli/test/passthrough.test.js
@@ -18,153 +18,6 @@ afterEach(() => {
 });
 
 describe('Wiggum CLI Passthrough Tests', () => {
-  describe('--version flag passthrough', () => {
-    test('pack --version should return rspack version', () => {
-      const result = runCLI('pack --version');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout.trim()).toMatch(SEMVER_OR_RSPACK_VERSION);
-    });
-
-    test('doc --version should return rspress version', () => {
-      const result = runCLI('doc --version');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('Rspress');
-      expect(result.stdout).toMatch(/\d+\.\d+\.\d+/);
-    });
-
-    test('lib --version should return rslib version', () => {
-      const result = runCLI('lib --version');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('Rslib');
-      expect(result.stdout).toMatch(/\d+\.\d+\.\d+/);
-    });
-
-    test('test --version should return rstest version', () => {
-      const result = runCLI('test --version');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rstest');
-      expect(result.stdout).toMatch(/\d+\.\d+\.\d+/);
-    });
-
-    test('build --version should return rsbuild version', () => {
-      const result = runCLI('build --version');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('Rsbuild');
-      expect(result.stdout).toMatch(/\d+\.\d+\.\d+/);
-    });
-  });
-
-  describe('--help flag passthrough', () => {
-    test('pack --help should return rspack help', () => {
-      const result = runCLI('pack --help');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rspack');
-      expect(result.stdout).toContain('Commands:');
-      expect(result.stdout).toContain('Options:');
-    });
-
-    test('doc --help should return rspress help', () => {
-      const result = runCLI('doc --help');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rspress');
-      expect(result.stdout).toContain('Usage:');
-    });
-
-    test('build --help should return rsbuild help', () => {
-      const result = runCLI('build --help');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rsbuild');
-      expect(result.stdout).toContain('Usage:');
-    });
-
-    test('lib --help should return rslib help', () => {
-      const result = runCLI('lib --help');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rslib');
-      expect(result.stdout).toContain('Usage:');
-    });
-
-    test('test --help should return rstest help', () => {
-      const result = runCLI('test --help');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rstest');
-      expect(result.stdout).toContain('Usage:');
-    });
-  });
-
-  describe('-v flag passthrough (short version)', () => {
-    test('pack -v should return rspack version', () => {
-      const result = runCLI('pack -v');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout.trim()).toMatch(SEMVER_OR_RSPACK_VERSION);
-    });
-
-    test('doc -v should return rspress version', () => {
-      const result = runCLI('doc -v');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('Rspress');
-      expect(result.stdout).toMatch(/\d+\.\d+\.\d+/);
-    });
-
-    test('lib -v should return rslib version', () => {
-      const result = runCLI('lib -v');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('Rslib');
-      expect(result.stdout).toMatch(/\d+\.\d+\.\d+/);
-    });
-
-    test('test -v should return rstest version', () => {
-      const result = runCLI('test -v');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rstest');
-      expect(result.stdout).toMatch(/\d+\.\d+\.\d+/);
-    });
-
-    test('build -v should return rsbuild version', () => {
-      const result = runCLI('build -v');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('Rsbuild');
-      expect(result.stdout).toMatch(/\d+\.\d+\.\d+/);
-    });
-  });
-
-  describe('-h flag passthrough (short help)', () => {
-    test('pack -h should return rspack help', () => {
-      const result = runCLI('pack -h');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rspack');
-      expect(result.stdout).toContain('Commands:');
-    });
-
-    test('doc -h should return rspress help', () => {
-      const result = runCLI('doc -h');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rspress');
-      expect(result.stdout).toContain('Usage:');
-    });
-
-    test('lib -h should return rslib help', () => {
-      const result = runCLI('lib -h');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rslib');
-      expect(result.stdout).toContain('Usage:');
-    });
-
-    test('test -h should return rstest help', () => {
-      const result = runCLI('test -h');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rstest');
-      expect(result.stdout).toContain('Usage:');
-    });
-
-    test('build -h should return rsbuild help', () => {
-      const result = runCLI('build -h');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rsbuild');
-      expect(result.stdout).toContain('Usage:');
-    });
-  });
-
   describe('Wiggum-specific flags', () => {
     test('--help should show wiggum help', () => {
       const result = runCLI('--help');
@@ -1005,166 +858,4 @@ describe('Wiggum CLI Passthrough Tests', () => {
     });
   });
 
-  describe('Complex flag combinations', () => {
-    test('build --mode production --help should forward all flags', () => {
-      const result = runCLI('build --mode production --help');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('rsbuild');
-      expect(result.stdout).toContain('Usage:');
-    });
-
-    test('pack --config webpack.config.js --version should forward all flags', () => {
-      const result = runCLI('pack --config webpack.config.js --version');
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout.trim()).toMatch(SEMVER_OR_RSPACK_VERSION);
-    });
-
-    test('build forwards --autofix when passed after delimiter', () => {
-      const root = makeTempDir();
-      const binDir = path.join(root, 'bin');
-      fs.mkdirSync(binDir, { recursive: true });
-      const fakeRsbuildPath = path.join(binDir, 'rsbuild');
-      fs.writeFileSync(
-        fakeRsbuildPath,
-        '#!/usr/bin/env bash\necho \"fake-rsbuild:$@\"\nexit 0\n',
-        { mode: 0o755 },
-      );
-      fs.chmodSync(fakeRsbuildPath, 0o755);
-
-      const result = runCLI('build -- --autofix', {
-        cwd: root,
-        env: {
-          ...process.env,
-          PATH: `${binDir}:${process.env.PATH || ''}`,
-        },
-      });
-
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain('fake-rsbuild:-- --autofix');
-    });
-
-    test('global --autofix triggers prompt flow for passthrough failures', () => {
-      const root = makeTempDir();
-      const binDir = path.join(root, 'bin');
-      fs.mkdirSync(binDir, { recursive: true });
-      const fakeRsbuildPath = path.join(binDir, 'rsbuild');
-      fs.writeFileSync(
-        fakeRsbuildPath,
-        '#!/usr/bin/env bash\necho \"pass stdout\"\necho \"pass stderr\" 1>&2\nexit 2\n',
-        { mode: 0o755 },
-      );
-      fs.chmodSync(fakeRsbuildPath, 0o755);
-
-      const result = runCLI('build --autofix', {
-        cwd: root,
-        env: {
-          ...process.env,
-          PATH: `${binDir}:${process.env.PATH || ''}`,
-          WIGGUM_AUTOFIX_MODE: 'prompt',
-        },
-      });
-
-      expect(result.exitCode).toBe(1);
-      expect(result.stdout).toContain('[autofix] Prompt-only mode enabled.');
-      expect(result.stdout).toContain('Command failed: wiggum rsbuild');
-      expect(result.stdout).toContain('pass stdout');
-      expect(result.stdout).toContain('pass stderr');
-    });
-
-    test('leading global --autofix triggers prompt flow for passthrough failures', () => {
-      const root = makeTempDir();
-      const binDir = path.join(root, 'bin');
-      fs.mkdirSync(binDir, { recursive: true });
-      const fakeRsbuildPath = path.join(binDir, 'rsbuild');
-      fs.writeFileSync(
-        fakeRsbuildPath,
-        '#!/usr/bin/env bash\necho \"pass stdout\"\necho \"pass stderr\" 1>&2\nexit 2\n',
-        { mode: 0o755 },
-      );
-      fs.chmodSync(fakeRsbuildPath, 0o755);
-
-      const result = runCLI('--autofix build', {
-        cwd: root,
-        env: {
-          ...process.env,
-          PATH: `${binDir}:${process.env.PATH || ''}`,
-          WIGGUM_AUTOFIX_MODE: 'prompt',
-        },
-      });
-
-      expect(result.exitCode).toBe(1);
-      expect(result.stdout).toContain('[autofix] Prompt-only mode enabled.');
-      expect(result.stdout).toContain('Command failed: wiggum rsbuild');
-      expect(result.stdout).toContain('pass stdout');
-      expect(result.stdout).toContain('pass stderr');
-    });
-
-    test('leading global --autofix keeps delimiter-passed tool args', () => {
-      const root = makeTempDir();
-      const binDir = path.join(root, 'bin');
-      fs.mkdirSync(binDir, { recursive: true });
-      const fakeRsbuildPath = path.join(binDir, 'rsbuild');
-      fs.writeFileSync(
-        fakeRsbuildPath,
-        '#!/usr/bin/env bash\necho \"pass stdout\"\necho \"pass stderr\" 1>&2\nexit 2\n',
-        { mode: 0o755 },
-      );
-      fs.chmodSync(fakeRsbuildPath, 0o755);
-
-      const result = runCLI('--autofix build -- --autofix', {
-        cwd: root,
-        env: {
-          ...process.env,
-          PATH: `${binDir}:${process.env.PATH || ''}`,
-          WIGGUM_AUTOFIX_MODE: 'prompt',
-        },
-      });
-
-      expect(result.exitCode).toBe(1);
-      expect(result.stdout).toContain('[autofix] Prompt-only mode enabled.');
-      expect(result.stdout).toContain('Command failed: wiggum rsbuild -- --autofix');
-      expect(result.stdout).toContain('pass stdout');
-      expect(result.stdout).toContain('pass stderr');
-    });
-
-    test('inline global --autofix keeps delimiter-passed tool args', () => {
-      const root = makeTempDir();
-      const binDir = path.join(root, 'bin');
-      fs.mkdirSync(binDir, { recursive: true });
-      const fakeRsbuildPath = path.join(binDir, 'rsbuild');
-      fs.writeFileSync(
-        fakeRsbuildPath,
-        '#!/usr/bin/env bash\necho \"pass stdout\"\necho \"pass stderr\" 1>&2\nexit 2\n',
-        { mode: 0o755 },
-      );
-      fs.chmodSync(fakeRsbuildPath, 0o755);
-
-      const result = runCLI('build --autofix -- --autofix', {
-        cwd: root,
-        env: {
-          ...process.env,
-          PATH: `${binDir}:${process.env.PATH || ''}`,
-          WIGGUM_AUTOFIX_MODE: 'prompt',
-        },
-      });
-
-      expect(result.exitCode).toBe(1);
-      expect(result.stdout).toContain('[autofix] Prompt-only mode enabled.');
-      expect(result.stdout).toContain('Command failed: wiggum rsbuild -- --autofix');
-      expect(result.stdout).toContain('pass stdout');
-      expect(result.stdout).toContain('pass stderr');
-    });
-
-    test('should handle invalid commands gracefully', () => {
-      const result = runCLI('nonexistent --version');
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toContain('Unknown command: nonexistent');
-    });
-
-    test('should handle invalid commands with leading global --autofix gracefully', () => {
-      const result = runCLI('--autofix nonexistent --version');
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toContain('Unknown command: nonexistent');
-    });
-  });
 });
